+++ packages/expo-updates/CHANGELOG.md
@@ -11,6 +11,7 @@
 ### 🎉 New features
 
 - Add rollback to embedded update directive. ([#21007](https://github.com/expo/expo/pull/21007), [#21652](https://github.com/expo/expo/pull/21652) by [@wschurman](https://github.com/wschurman))
+- Add support for extra params. ([#21837](https://github.com/expo/expo/pull/21837) by [@wschurman](https://github.com/wschurman))
 
 ### 🐛 Bug fixes